rzutowanie v na double jest nadmiarowe

+++ MasterCoder.PKW.Mandates/MandateMethodDHondta.cs
@@ -49,7 +49,8 @@ namespace MasterCoder.PKW.Mandates
 
         private double Calc(int v, int mandates)
         {
-            return ((double)v * 1.0 / (2*mandates + 1.0));
+            //rzutowanie v na double jest nadmiarowe
+            return (v * 1.0 / (2*mandates + 1.0));
         }
     }
 }
